* Removed some warnings
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8410 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/ltsview/mathutils.h
@@ -14,7 +14,7 @@
 
 namespace MathUtils
 {
-  const float PI = 3.14159265359;
+  const double PI = 3.14159265359;
   int round_to_int(float f);
   float deg_to_rad(float deg);
   float rad_to_deg(float rad);
